pyproject.toml
[project]
name = "beyond-the-cutoff"
version = "0.1.0"
description = "Evaluation of LLM knowledge updating beyond pretraining cutoffs"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Ignacio" }
]
dependencies = [
  "mlx>=0.0.8",
  "transformers>=4.44",
  "datasets>=2.21",
  "peft>=0.12",
  "accelerate>=0.30",
  "pypdf>=4.3",
  "pymupdf>=1.24",
  "faiss-cpu>=1.8",
  "chromadb>=0.5",
  "evaluate>=0.4",
  "bert-score>=0.3",
  "sentence-transformers>=3.0",
  "hydra-core>=1.3",
  "typer>=0.12",
  "pydantic>=2.8",
  "omegaconf>=2.3",
  "PyYAML>=6.0",
  "httpx>=0.27",
  "uvicorn>=0.30",
  "fastapi>=0.111",
  "numpy>=1.26",
  "scipy>=1.13",
  "scikit-learn>=1.5",
  "pandas>=2.2",
  "rich>=13.7",
  "matplotlib>=3.9",
  "seaborn>=0.13",
  "plotly>=5.23",
  "jupyter>=1.0",
  "ipykernel>=6.29",
  "pyarrow>=16.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.7",
  "mypy>=1.11",
  "pre-commit>=3.8",
  "types-PyYAML>=6.0.12"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
extend-select = [
  "B",
  "I",
  "N",
  "UP",
  "C4"
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
warn_return_any = true
no_implicit_optional = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
  "numpy",
  "numpy.*",
  "faiss",
  "faiss.*",
  "sentence_transformers",
  "sentence_transformers.*",
  "fitz",
  "fitz.*",
  "nltk",
  "nltk.*",
  "pytest",
  "pytest.*",
  "pypdf",
  "pypdf.*",
  "fastapi",
  "fastapi.*",
  "httpx",
  "httpx.*",
  "pandas",
  "pandas.*",
  "evaluate",
  "evaluate.*"
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
  "datasets",
  "datasets.*",
  "peft",
  "peft.*",
  "transformers",
  "transformers.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra --maxfail=1 --strict-markers"
testpaths = [
  "tests"
]
